Renames class to TwelveHourTime and adds namespacing
