+++ military_time_conversion/miltime_convert.rb
@@ -7,7 +7,7 @@ print "Input military time (xxxx): "
 begin
   # Get a number from stdin. Make it base-10 to avoid converting "0123" to octal
   military_time = Integer(gets, 10)
-  puts MilitaryTime.new(military_time)
+  puts MilitaryTimeConversion::TwelveHourTime.new(military_time)
 rescue ArgumentError
   puts "You've entered an invalid time."
 end
